+++ reader-gtfs/src/main/java/com/graphhopper/gtfs/TripFromLabel.java
@@ -342,9 +342,9 @@ class TripFromLabel {
         }
         if (GtfsStorage.EdgeType.ENTER_PT == path.get(1).edge.getType()) {
             String feedId = path.get(1).edge.getPlatformDescriptor().feed_id;
+            int routeType = path.get(1).edge.getRouteType();
             List<Trip.Leg> result = new ArrayList<>();
             long boardTime = -1;
-            int routeType = -1;
             List<Label.Transition> partition = null;
             for (int i = 1; i < path.size(); i++) {
                 Label.Transition transition = path.get(i);
@@ -352,8 +352,6 @@ class TripFromLabel {
                 if (edge.getType() == GtfsStorage.EdgeType.BOARD) {
                     boardTime = transition.label.currentTime;
                     partition = new ArrayList<>();
-                } else if (edge.getType() == GtfsStorage.EdgeType.ENTER_PT) {
-                    routeType = edge.getRouteType();
                 }
                 if (partition != null) {
                     partition.add(path.get(i));
@@ -367,10 +365,21 @@ class TripFromLabel {
                     stopsFromBoardHopDwellEdges.finish();
                     List<Trip.Stop> stops = stopsFromBoardHopDwellEdges.stops;
 
+                    GTFSFeed gtfsFeed = gtfsStorage.getGtfsFeeds().get(feedId);
+                    String routeId = tripDescriptor.getRouteId();
+                    String agencyId = gtfsFeed.routes.get(routeId).agency_id;
+                    String agencyName = (agencyId != null
+                                    ? Optional.ofNullable(gtfsFeed.agency.get(agencyId))
+                                    : gtfsFeed.agency.values().stream().findFirst()
+                            ).map(a -> a.agency_name).orElse("");
+
                     result.add(new Trip.PtLeg(
-                            feedId, partition.get(0).edge.getTransfers() == 0,
+                            feedId,
+                            agencyId,
+                            agencyName,
+                            partition.get(0).edge.getTransfers() == 0,
                             tripDescriptor.getTripId(),
-                            tripDescriptor.getRouteId(),
+                            routeId,
                             routeType,
                             Optional.ofNullable(gtfsStorage.getGtfsFeeds().get(feedId).trips.get(tripDescriptor.getTripId())).map(t -> t.trip_headsign).orElse("extra"),
                             stops,