+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - car.json by default avoids private roads
 - maxspeed<5 is ignored, maxspeed=none is ignored with some exceptions, maxspeed parsing and related constants were renamed #3077
 - improved performance by sorting graph during import, #3177
+- trunk roads in Austria are no longer considered to be toll roads by default
 
 ### 10.0 [5 Nov 2024]